+++ nn_dataflow/tests/loop_blocking_test/test_loop_blocking_fixture.py
@@ -58,15 +58,18 @@ class TestLoopBlockingFixture(unittest.TestCase):
         # Typical resource.
         self.resource['BASE'] = Resource(
             proc_region=proc_region, data_regions=data_regions,
-            dim_array=dim_array, size_gbuf=65536, size_regf=64)
+            dim_array=dim_array, size_gbuf=65536, size_regf=64,
+            array_bus_width=float('inf'))
         # Larger resource with sufficient capacity, to make all schemes valid.
         self.resource['LG'] = Resource(
             proc_region=proc_region, data_regions=data_regions,
-            dim_array=dim_array, size_gbuf=1024 ** 3, size_regf=1024 ** 3)
+            dim_array=dim_array, size_gbuf=1024 ** 3, size_regf=1024 ** 3,
+            array_bus_width=float('inf'))
         # Small resource.
         self.resource['SM'] = Resource(
             proc_region=proc_region, data_regions=data_regions,
-            dim_array=dim_array, size_gbuf=4096, size_regf=16)
+            dim_array=dim_array, size_gbuf=4096, size_regf=16,
+            array_bus_width=float('inf'))
 
         # Nested loop description after mapping.
         self.nld = {}