+++ nn_dataflow/tests/unit_test/test_resource.py
@@ -41,11 +41,13 @@ class TestResource(unittest.TestCase):
                             dim_array=PhyDim2(16, 16),
                             size_gbuf=131072,
                             size_regf=512,
+                            array_bus_width=8,
                            )
         self.assertTupleEqual(resource.proc_region.dim, (2, 2), 'proc_region')
         self.assertTupleEqual(resource.dim_array, (16, 16), 'dim_array')
         self.assertEqual(resource.size_gbuf, 131072, 'size_gbuf')
         self.assertEqual(resource.size_regf, 512, 'size_regf')
+        self.assertEqual(resource.array_bus_width, 8, 'array_bus_width')
 
     def test_invalid_proc_region(self):
         ''' Invalid proc_region. '''
@@ -60,6 +62,7 @@ class TestResource(unittest.TestCase):
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=512,
+                         array_bus_width=8,
                         )
 
     def test_invalid_proc_region_data(self):
@@ -77,6 +80,7 @@ class TestResource(unittest.TestCase):
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=512,
+                         array_bus_width=8,
                         )
 
     def test_invalid_dim_array(self):
@@ -94,6 +98,7 @@ class TestResource(unittest.TestCase):
                          dim_array=(16, 16),
                          size_gbuf=131072,
                          size_regf=512,
+                         array_bus_width=8,
                         )
 
     def test_invalid_size_gbuf(self):
@@ -111,6 +116,7 @@ class TestResource(unittest.TestCase):
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=(131072,),
                          size_regf=512,
+                         array_bus_width=8,
                         )
 
     def test_invalid_size_regf(self):
@@ -128,6 +134,58 @@ class TestResource(unittest.TestCase):
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=(512,),
+                         array_bus_width=8,
+                        )
+
+    def test_invalid_array_bus_width(self):
+        ''' Invalid array_bus_width. '''
+        with self.assertRaisesRegexp(TypeError,
+                                     'Resource: .*array_bus_width.*'):
+            _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0),
+                                                type=NodeRegion.PROC),
+                         data_regions=(NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 0),
+                                                  type=NodeRegion.DATA),
+                                       NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 1),
+                                                  type=NodeRegion.DATA)),
+                         dim_array=PhyDim2(16, 16),
+                         size_gbuf=131072,
+                         size_regf=512,
+                         array_bus_width=1.2,
+                        )
+        with self.assertRaisesRegexp(ValueError,
+                                     'Resource: .*array_bus_width.*'):
+            _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0),
+                                                type=NodeRegion.PROC),
+                         data_regions=(NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 0),
+                                                  type=NodeRegion.DATA),
+                                       NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 1),
+                                                  type=NodeRegion.DATA)),
+                         dim_array=PhyDim2(16, 16),
+                         size_gbuf=131072,
+                         size_regf=512,
+                         array_bus_width=-2,
+                        )
+        with self.assertRaisesRegexp(ValueError,
+                                     'Resource: .*array_bus_width.*'):
+            _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0),
+                                                type=NodeRegion.PROC),
+                         data_regions=(NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 0),
+                                                  type=NodeRegion.DATA),
+                                       NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 1),
+                                                  type=NodeRegion.DATA)),
+                         dim_array=PhyDim2(16, 16),
+                         size_gbuf=131072,
+                         size_regf=512,
+                         array_bus_width=0,
                         )
 
     def test_invalid_data_regions_type(self):
@@ -142,6 +200,7 @@ class TestResource(unittest.TestCase):
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=(512,),
+                         array_bus_width=8,
                         )
         with self.assertRaisesRegexp(TypeError, 'Resource: .*data_regions.*'):
             _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
@@ -153,6 +212,7 @@ class TestResource(unittest.TestCase):
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=(512,),
+                         array_bus_width=8,
                         )
         with self.assertRaisesRegexp(TypeError, 'Resource: .*data_regions.*'):
             _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
@@ -165,6 +225,7 @@ class TestResource(unittest.TestCase):
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=(512,),
+                         array_bus_width=8,
                         )
 
     def test_invalid_data_regions_len(self):
@@ -179,6 +240,7 @@ class TestResource(unittest.TestCase):
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=(512,),
+                         array_bus_width=8,
                         )
 
     def test_invalid_data_regions_proc(self):
@@ -196,6 +258,7 @@ class TestResource(unittest.TestCase):
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=512,
+                         array_bus_width=8,
                         )
 
     def test_src_data_region(self):
@@ -210,6 +273,7 @@ class TestResource(unittest.TestCase):
                             dim_array=PhyDim2(16, 16),
                             size_gbuf=131072,
                             size_regf=512,
+                            array_bus_width=8,
                             data_regions=(nr1, nr2))
         self.assertEqual(resource.src_data_region(), nr1, 'src_data_region')
         resource = Resource(proc_region=NodeRegion(dim=PhyDim2(4, 4),
@@ -218,6 +282,7 @@ class TestResource(unittest.TestCase):
                             dim_array=PhyDim2(16, 16),
                             size_gbuf=131072,
                             size_regf=512,
+                            array_bus_width=8,
                             data_regions=(nr1,))
         self.assertEqual(resource.src_data_region(), nr1, 'src_data_region')
 
@@ -233,6 +298,7 @@ class TestResource(unittest.TestCase):
                             dim_array=PhyDim2(16, 16),
                             size_gbuf=131072,
                             size_regf=512,
+                            array_bus_width=8,
                             data_regions=(nr1, nr2))
         self.assertEqual(resource.dst_data_region(), nr2, 'dst_data_region')
         resource = Resource(proc_region=NodeRegion(dim=PhyDim2(4, 4),
@@ -241,6 +307,7 @@ class TestResource(unittest.TestCase):
                             dim_array=PhyDim2(16, 16),
                             size_gbuf=131072,
                             size_regf=512,
+                            array_bus_width=8,
                             data_regions=(nr1,))
         self.assertEqual(resource.dst_data_region(), nr1, 'dst_data_region')