+++ nn_dataflow/tools/nn_dataflow_search.py
@@ -101,6 +101,10 @@ def do_scheduling(args):
     size_gbuf = args.gbuf / word
     size_regf = args.regf / word
 
+    array_bus_width = args.bus_width // args.word
+    if not array_bus_width:
+        array_bus_width = float('inf')
+
     proc_region = NodeRegion(dim=dim_nodes,
                              origin=PhyDim2(0, 0),
                              type=NodeRegion.PROC)
@@ -123,7 +127,8 @@ def do_scheduling(args):
                         data_regions=data_regions,
                         dim_array=dim_array,
                         size_gbuf=size_gbuf,
-                        size_regf=size_regf)
+                        size_regf=size_regf,
+                        array_bus_width=array_bus_width)
 
     ## Cost.
 
@@ -215,6 +220,9 @@ def argparser():
                     help='register file size in bytes per PE')
     ap.add_argument('--gbuf', type=int, required=True,
                     help='global buffer size in bytes')
+
+    ap.add_argument('--bus-width', type=int, default=0,
+                    help='array bus width in bits. set 0 to ignore')
 
     ap.add_argument('--op-cost', type=float, default=1,
                     help='cost of arithmetic operation')