+++ CHANGELOG.md
@@ -4,7 +4,7 @@ List of major changes and improvements
 ## [Unreleased]
 
 
-## [v1.5 -- v1.6] -- 2017-10-26
+## [v1.5 -- v1.6] -- 2017-12-05
 
 ### Added